Fix i18n regressions
At least qubes-manager notifications (and likely other uses of
translated strings) have been broken since 4946d7b because .tr() is
attempted to be called on self in places where self is not the
QApplication.

This results in stack traces like the following: (when Ctrl-Shift-C'ing)
```
Exception in thread Thread-2:
Traceback (most recent call last):
  File "/usr/lib64/python2.7/threading.py", line 804, in __bootstrap_inner
    self.run()
  File "/usr/lib/python2.7/site-packages/pyinotify.py", line 1505, in run
    self.loop()
  File "/usr/lib/python2.7/site-packages/pyinotify.py", line 1491, in loop
    self.process_events()
  File "/usr/lib/python2.7/site-packages/pyinotify.py", line 1287, in process_events
    self._default_proc_fun(revent)
  File "/usr/lib/python2.7/site-packages/pyinotify.py", line 924, in __call__
    return _ProcessEvent.__call__(self, event)
  File "/usr/lib/python2.7/site-packages/pyinotify.py", line 644, in __call__
    return meth(event)
  File "/usr/lib64/python2.7/site-packages/qubesmanager/main.py", line 105, in process_IN_CLOSE_WRITE
    trayIcon.showMessage(unicode(self.tr(
AttributeError: 'QubesManagerFileWatcher' object has no attribute 'tr'
```

So instead, try calling .tr() on the (always?) globally available app.
I'm not sure if this is the correct thing to do, but it is more
reliable than calling on self.

This patch is just the result of:

    `find . -type f -exec sed -i 's/self\.tr(/app.tr(/' {} +`